# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StackOverStay Mobile API",
    "version": "1.0.0",
    "description": "API documentation for the StackOverStay mobile application backend"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local server"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "StandardResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "type": "object"
          },
          "meta": {
            "type": "object"
          }
        }
      }
    }
  },
  "paths": {
    "/banners": {
      "get": {
        "summary": "Get list of banners",
        "tags": [
          "Banners"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "platform",
            "schema": {
              "type": "string",
              "default": "mobile"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of banners",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "title": {
                                "type": "string"
                              },
                              "imageUrl": {
                                "type": "string"
                              },
                              "jump": {
                                "type": "object",
                                "properties": {
                                  "type": {
                                    "type": "string"
                                  },
                                  "payload": {
                                    "type": "object"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/hotels": {
      "get": {
        "summary": "Search hotels",
        "tags": [
          "Hotels"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "cityId",
            "schema": {
              "type": "integer"
            },
            "required": false,
            "description": "City ID (required if no keyword)"
          },
          {
            "in": "query",
            "name": "keyword",
            "schema": {
              "type": "string"
            },
            "description": "Search keyword"
          },
          {
            "in": "query",
            "name": "checkIn",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Check-in date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "checkOut",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Check-out date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of hotels",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "name": {
                                "type": "string"
                              },
                              "starRating": {
                                "type": "number"
                              },
                              "minPrice": {
                                "type": "number"
                              }
                            }
                          }
                        },
                        "meta": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "integer"
                            },
                            "page": {
                              "type": "integer"
                            },
                            "pageSize": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/hotels/{hotelId}": {
      "get": {
        "summary": "Get hotel details",
        "tags": [
          "Hotels"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "hotelId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel details",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer"
                            },
                            "name": {
                              "type": "string"
                            },
                            "description": {
                              "type": "string"
                            },
                            "photos": {
                              "type": "array",
                              "items": {
                                "type": "object"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Hotel not found"
          }
        }
      }
    },
    "/hotels/{hotelId}/rooms": {
      "get": {
        "summary": "Get hotel rooms and rates",
        "tags": [
          "Hotels"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "hotelId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "checkIn",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "checkOut",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of rooms with pricing"
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check endpoint",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "Server is up and running",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/meta/cities": {
      "get": {
        "summary": "Get list of cities",
        "tags": [
          "Meta"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "schema": {
              "type": "string"
            },
            "description": "Search by city name"
          },
          {
            "in": "query",
            "name": "country",
            "schema": {
              "type": "string"
            },
            "description": "Filter by country code"
          }
        ],
        "responses": {
          "200": {
            "description": "List of cities",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "name": {
                                "type": "string"
                              },
                              "countryCode": {
                                "type": "string"
                              },
                              "lat": {
                                "type": "number"
                              },
                              "lng": {
                                "type": "number"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/meta/quick-filters": {
      "get": {
        "summary": "Get quick filters configuration",
        "tags": [
          "Meta"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "platform",
            "schema": {
              "type": "string",
              "default": "mobile"
            },
            "description": "Platform identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "List of quick filters"
          }
        }
      }
    },
    "/meta/hotel-filters": {
      "get": {
        "summary": "Get hotel filter options",
        "tags": [
          "Meta"
        ],
        "responses": {
          "200": {
            "description": "Hotel filter options"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Banners",
      "description": "Banner management"
    },
    {
      "name": "Hotels",
      "description": "Hotel search and details"
    },
    {
      "name": "Meta",
      "description": "Metadata and configuration endpoints"
    }
  ]
}